Add hydration example
