--- examples/hello-hydrate/index.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <title>Hello World!</title>
-
-    <style>
-      body {
-        font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
-      }
-    </style>
-  </head>
-  <body>
-    <!-- SSR-ed HTML can be found in dev console when running the app. -->
-    <div data-hk="1.0">
-      <h1 data-hk="1.1">Hello <span data-hk="1.2">World</span>!</h1>
-      <input data-hk="1.3" />
-    </div>
-  </body>
-</html>
